4. ImageChat/requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 